Split a large source file into smaller ones
